+++ CHANGELOG.md
@@ -61,6 +61,8 @@
   specified.)
   ([#95](https://github.com/biocore/qurro/issues/95))
 ### Backward-incompatible changes
+- Qurro will now try to load the `Differential` type from q2-types. **This means
+  that Qurro v0.3.0 will only support versions of QIIME 2 of at least 2019.7.**
 - Renamed Qurro's QIIME 2 actions ([#98](https://github.com/biocore/qurro/issues/98)):
     - **Renamed `qiime qurro supervised-rank-plot` to `qiime qurro differential-plot`.**
     - **Renamed `qiime qurro unsupervised-rank-plot` to `qiime qurro loading-plot`.**